Add files via upload
An individual project through which I've attempted to incorporate a the topic of recycling and make it interactive/appealing for the user. This project thought me about UI and UX. I bumped into an issue, regarding delay, when trying to load images from my data library but I managed to overcome this by renaming each image. I used a structure close to an array for the loading of the images by giving them all the same primary name (just like an array of the same variables) and gave them different values. This managed to decrease the run time of my application by 25% which overcame lag and ensured reliability.